Light square on client side in case web socket connects then fails
Currently the websocket-enabled lighting of squares waits for a
websocket notification rather than fake lighting a square on the field
first and then confirming. This leads to blank squares displaying to the
user if their websocket connection fails in some way. This fixes things
by fake lighting the square on the client side and then updating it
again when the true status is set via websockets.